hardware_info/requirements.txt
psutil
py-cpuinfo
wmi
tabulate